Package References

Package References

General function

Algorithm

General function

ImageSmooth.SmoothAPI.smooth — Function

smooth(img, fₛ::AbstractImageSmoothAlgorithm, args...; kwargs...)

Smooth img using algorithm fₛ

Output

The return image imgₛ is an Array{Gray{Float64}} forGrayimage, andArray{RGB{Float64}}forRGB` image.

Examples

Just simply pass the input image and algorithm to smooth

fₛ = L0Smooth()

imgₛ = smooth(img, fₛ)

This reads as "smooth image img using smoothing algorithm fₛ".

See also smooth! for in-place smoothing.

ImageSmooth.SmoothAPI.smooth! — Function

smooth!(out::AbstractArray{<: Number}, img::AbstractArray{<: Number}, fₛ::AbstractImageSmoothAlgorithm, args...; kwargs...)

Smooth img::AbstractArray{<: Number} using algorithm fₛ

Output

out will be changed in place.

Examples

Just simply pass an algorithm to smooth!:

# First generate an algorithm instance
fₛ = L0Smooth()

## For Gray images
input = reshape(channelview(img), 1, size(img)...)
imgₛ = similar(float64.(input))

smooth!(imgₛ, input, fₛ)

## For RGB images
input = channelview(img)
imgₛ = similar(float64.(input))

smooth!(imgₛ, img, fₛ)

See also: smooth

ImageSmooth.forwarddiff — Function

forwarddiff(a::AbstractArray{T,N}; dims::Int) where {T,N}

Finite one-dimension forward difference operator on a vector or a multidimensional array A. In both cases, the dimension to operate on needs to be specified with the dims keyword argument.

Output

The return Vector or Array maintains the same size as input.

Examples

julia> A = [2 4 8; 3 9 27; 4 16 64]
3×3 Matrix{Int64}:
 2   4   8
 3   9  27
 4  16  64

julia> forwarddiff(A, dims=2)
3×3 Matrix{Int64}:
  2   4   -6
  6  18  -24
 12  48  -60

See also forwarddiff! for in-place forward difference.

ImageSmooth.forwarddiff! — Function

forwarddiff!(d::AbstractArray{T,N}, a::AbstractArray{T,N}; dims::Int) where {T,N}}

Finite one-dimension forward difference operator on a vector or a multidimensional array A. In both cases, the dimension to operate on needs to be specified with the dims keyword argument.

Output

d will be changed in place. The return Vector or Array maintains the same size as input.

Examples

julia> A = [2 4 8; 3 9 27; 4 16 64]
3×3 Matrix{Int64}:
 2   4   8
 3   9  27
 4  16  64

julia> D = similar(A)
3×3 Matrix{Int64}:
 140441636044112  140441636135216  140441696235920
 140441647064848  140441636135296  140441636135456
 140441636135136  140441636135376  140441636135056

julia> forwarddiff!(D, A, dims=2)
3×3 Matrix{Int64}:
  2   4   -6
  6  18  -24
 12  48  -60

julia> D
3×3 Matrix{Int64}:
  2   4   -6
  6  18  -24
 12  48  -60

See also: forwarddiff

ImageSmooth.backdiff — Function

backdiff(a::AbstractArray{T,N}; dims::Int) where {T,N}

Finite one-dimension backward difference operator on a vector or a multidimensional array A. In both cases, the dimension to operate on needs to be specified with the dims keyword argument.

Output

The return Vector or Array maintains the same size as input.

Examples

julia> A = [2 4 8; 3 9 27; 4 16 64]
3×3 Matrix{Int64}:
 2   4   8
 3   9  27
 4  16  64

julia> backdiff(A, dims=2)
3×3 Matrix{Int64}:
  6   -2   -4
 24   -6  -18
 60  -12  -48

See also backdiff! for in-place backward difference.

ImageSmooth.backdiff! — Function

backdiff!(d::AbstractArray{T,N}, a::AbstractArray{T,N}; dims::Int) where {T,N}}

Finite one-dimension backward difference operator on a vector or a multidimensional array A. In both cases, the dimension to operate on needs to be specified with the dims keyword argument.

Output

d will be changed in place. The return Vector or Array maintains the same size as input.

Examples

julia> A = [2 4 8; 3 9 27; 4 16 64]
3×3 Matrix{Int64}:
 2   4   8
 3   9  27
 4  16  64

julia> D = similar(A)
3×3 Matrix{Int64}:
 140103024112336  140103024112336  140103027568240
               0                0                0
 140103024361232  140103024361232  140103024361232

julia> backdiff!(D, A, dims=2)
3×3 Matrix{Int64}:
  6   -2   -4
 24   -6  -18
 60  -12  -48

julia> D
3×3 Matrix{Int64}:
  6   -2   -4
 24   -6  -18
 60  -12  -48

See also: backdiff

Algorithm

ImageSmooth.SmoothAPI.AbstractImageAlgorithm — Type

AbstractImageAlgorithm

The root of image algorithms type system

ImageSmooth.SmoothAPI.AbstractImageFilter — Type

AbstractImageFilter <: AbstractImageAlgorithm

Filters are image algorithms whose input and output are both images

ImageSmooth.SmoothAPI.AbstractImageSmoothAlgorithm — Type

AbstractImageSmoothAlgorithm <: AbstractImageFilter

The root type for ImageSmooth package.

Any image smoothing algorithm shall subtype it to support smooth and smooth! APIs.

Examples

All algorithm in ImageSmooth are called in the following pattern, take L0Smooth <: AbstractImageSmoothAlgorithm as an example:

# First generate an algorithm instance
fₛ = L0Smooth()

# Then pass the algorithm to `smooth`
imgₛ = smooth(img, fₛ)

# Or use in-place version `smooth!`

## Ror Gray images
input = reshape(channelview(img), 1, size(img)...)
imgₛ = similar(float64.(input))

smooth!(imgₛ, input, fₛ)

## For RGB images
input = channelview(img)
imgₛ = similar(float64.(input))

smooth!(imgₛ, img, fₛ)

Some algorithms also receive parameters to control the smoothing process and to get an expected smooth degree.

# You could explicit specify it
f = L0Smooth(λ=0.04, κ=1.5, βmax=1e6)

# Or infer the default value
f = L0Smooth()

For more examples, please check smooth and smooth! and concret algorithms.